Combine view and controller, then split up into smaller packages
