+++ pom.xml
@@ -13,7 +13,7 @@
 
         <resources>
             <resource>
-                <directory>src/main/java/storagesystem/View</directory>
+                <directory>src/main/java/storagesystem/viewcontroller</directory>
             </resource>
             <resource>
                 <directory>src/main/resources</directory>
