+++ src/main/java/storagesystem/viewcontroller/ILoadUI.java
@@ -1,7 +1,8 @@
-package storagesystem.controller;
+package storagesystem.viewcontroller;
 
 /**
  * An interface for controllers with one method to avoid code repetition.
+ * @author William Albertsson
  */
 public interface ILoadUI {
