+++ src/main/java/storagesystem/viewcontroller/basicStyles.css
@@ -1,7 +1,3 @@
-/**
-
- */
-
 * {
     primaryColor: #424242;
     secondaryColor: #212121;
@@ -18,7 +14,6 @@
     -fx-font-size: 20;
 
 }
-
 
 .background {
     -fx-background-color: #303030;
