+++ src/main/java/storagesystem/viewcontroller/framework.fxml
@@ -15,7 +15,7 @@
 
 <AnchorPane fx:id="rootPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
             prefHeight="800.0" prefWidth="1200.0" xmlns="http://javafx.com/javafx/8.0.171"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.controller.FrameworkController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.viewcontroller.FrameworkController">
     <children>
         <BorderPane fx:id="borderPane" prefHeight="800.0" prefWidth="1200.0" styleClass="background"
                     stylesheets="@basicStyles.css">
