+++ src/main/java/storagesystem/viewcontroller/itemview/ItemPageController.java
@@ -1,8 +1,7 @@
-package storagesystem.controller;
+package storagesystem.viewcontroller.itemview;
 
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.Slider;
@@ -14,13 +13,12 @@ import storagesystem.model.Item;
 import storagesystem.model.Team;
 
 import java.io.IOException;
-import java.net.URL;
-import java.util.ResourceBundle;
 
 /**
  * Controls a detailed view of an item. Can be used to book an item.
  *
  * @author Jonathan Eksberg, Carl Lindh
+ * @revised by Hugo Stegrell
  */
 public class ItemPageController {
