+++ src/main/java/storagesystem/viewcontroller/itemview/itemList.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.FlowPane?>
 
 <AnchorPane fx:id="rootPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.controller.ItemListController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.viewcontroller.itemview.ItemListController">
     <children>
         <ScrollPane fx:id="itemListScrollPane" prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0"
                     AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
