+++ src/main/java/storagesystem/viewcontroller/itemview/itemPage.fxml
@@ -13,7 +13,7 @@
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
             prefWidth="700.0" style="-fx-background-color: rgba(0, 0, 0, 0.75);"
             xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="storagesystem.controller.ItemPageController">
+            fx:controller="storagesystem.viewcontroller.itemview.ItemPageController">
     <children>
         <AnchorPane layoutX="19.0" layoutY="30.0" style="-fx-background-color: white;" AnchorPane.leftAnchor="30.0"
                     AnchorPane.rightAnchor="30.0">
@@ -37,7 +37,7 @@
                 <ImageView fx:id="itemPageImageView" fitHeight="150.0" fitWidth="200.0" layoutY="5.0"
                            pickOnBounds="true" preserveRatio="true">
                     <image>
-                        <Image url="@../../../resources/art.png"/>
+                        <Image url="@../../../../resources/pictures/art.png"/>
                     </image>
                 </ImageView>
                 <Button fx:id="itemPageReserveBtn" layoutX="548.0" layoutY="492.0" mnemonicParsing="false"
