+++ src/main/java/storagesystem/viewcontroller/itemview/smallItemPanel.fxml
@@ -16,7 +16,7 @@
                    mouseTransparent="true" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="10.0"
                    AnchorPane.rightAnchor="10.0" AnchorPane.topAnchor="10.0">
             <image>
-                <Image url="@../../../resources/art.png"/>
+                <Image url="@../../../../resources/pictures/art.png"/>
             </image>
         </ImageView>
         <Label fx:id="itemNameLabel" alignment="CENTER" layoutX="73.0" layoutY="170.0" mouseTransparent="true"
