+++ src/main/java/storagesystem/viewcontroller/login/loginPage.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Text?>
 
 <AnchorPane fx:id="rootPane" prefHeight="800.0" prefWidth="1200.0" xmlns="http://javafx.com/javafx/8.0.171"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.controller.LoginPageController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="storagesystem.viewcontroller.login.LoginPageController">
     <children>
         <AnchorPane fx:id="leftAnchorPane" layoutY="171.0" prefHeight="800.0" prefWidth="600.0"
                     style="-fx-background-color: red;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0">
