+++ src/main/java/storagesystem/viewcontroller/reservations/ReservationDetailViewController.java
@@ -1,11 +1,10 @@
-package storagesystem.controller;
+package storagesystem.viewcontroller.reservations;
 
 import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
-import storagesystem.StoreIT;
 import storagesystem.model.IReservation;
 
 import java.io.IOException;
@@ -44,7 +43,7 @@ public class ReservationDetailViewController extends AnchorPane {
 
 
     public ReservationDetailViewController(IReservation res) {
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/reservationDetailView.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/reservations/reservationDetailView.fxml"));
         fxmlLoader.setController(this);
         fxmlLoader.setRoot(this);
