+++ src/main/java/storagesystem/viewcontroller/reservations/reservationDetailView.fxml
@@ -10,7 +10,7 @@
 
 
 <fx:root fx:id="rootPane" onMouseClicked="#closeReservationDetailView" prefHeight="700.0" prefWidth="1000.0"
-         styleClass="lightbox" stylesheets="@basicStyles.css" type="AnchorPane" xmlns="http://javafx.com/javafx/8.0.201"
+         styleClass="lightbox" stylesheets="@../basicStyles.css" type="AnchorPane" xmlns="http://javafx.com/javafx/8.0.201"
          xmlns:fx="http://javafx.com/fxml/1">
     <children>
         <AnchorPane fx:id="lightboxContentPane" blendMode="SRC_ATOP" styleClass="lightboxContent"
