+++ src/main/java/storagesystem/viewcontroller/reservations/reservations.fxml
@@ -11,8 +11,8 @@
 
 
 <AnchorPane fx:id="reservationsRootPane" prefHeight="700.0" prefWidth="1000.0" style="-fx-background-color: blue;"
-            stylesheets="@basicStyles.css" xmlns="http://javafx.com/javafx/8.0.201" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="storagesystem.controller.ReservationsController">
+            stylesheets="@../basicStyles.css" xmlns="http://javafx.com/javafx/8.0.201" xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="storagesystem.viewcontroller.reservations.ReservationsController">
     <children>
         <HBox alignment="TOP_CENTER" prefHeight="100.0" prefWidth="200.0" AnchorPane.leftAnchor="0.0"
               AnchorPane.rightAnchor="0.0">
