+++ src/main/java/storagesystem/viewcontroller/settings/SettingsController.java
@@ -1,4 +1,4 @@
-package storagesystem.controller;
+package storagesystem.viewcontroller.settings;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
