+++ src/main/resources/pictures/settingsIcon.png
